working on better checks

+++ src/controllers/user.controller.js
@@ -25,7 +25,7 @@ const registerUser = asyncHandler(async(req, res)=>{
     const existedUser = await User.findOne({email})
     
     if(existedUser){
-        
+        fs.unlinkSync(imageLocalPath);
         console.log(existedUser)
         return res.status(400).send({
             message:"User already exists",
@@ -34,21 +34,61 @@ const registerUser = asyncHandler(async(req, res)=>{
         
     }
     else{
-       const imageRes = await uploadOnCloudinary(imageLocalPath)
-       fs.unlinkSync(imageLocalPath)
-        const imageUrl = imageRes.url
-        const user = await User.create({email,password, imageUrl})
+        if(imageLocalPath){ 
+            console.log('image')
+            const imageRes = await uploadOnCloudinary(imageLocalPath)
+
+            fs.unlinkSync(imageLocalPath)
+            const imageUrl = imageRes.url
+            const user = await User.create({email,password, imageUrl})
+        }
+       
         
-        // console.log(user)
+     
+        else{
+            console.log('no  image')
+            const user = await User.create({email,password})
+        }
+        
         return res.status(201).send({
             message:"user created",
             success:true,
             user
-            // _id:user._id
         })
     }
 })
 
+const loginUser = asyncHandler(async(req, res)=>{
+    const {email,  password} = req.body;
+    if(!email){
+        return res.status(400).send({
+            message:"email is empty"
+        })
+    }
+    if(!password){
+        if(!password){
+    return res.status(400).send({
+        message:"password is empty"
+    })
+}
+    }
+
+    const user = await User.findOne({email});
+    if(user){
+        res.status(200).send({
+            user
+        })
+    }
+    else{
+        res.send(404).send({
+            message:"User does not exists"
+        })
+    }
+
+  
+
+})
 
 
-export {registerUser}
+
+export {registerUser, loginUser}
